# Conflicts:
#	src/main/java/app/rest/Routes.java

+++ src/main/java/app/daos/UserDAO.java
@@ -74,7 +74,7 @@ public class UserDAO
         return user;
     }
 
-    public User read(Integer id)
+    public User read(Long id)
     {
         try (EntityManager em = emf.createEntityManager())
         {
